Allow SSL configuration based on Puppet PEM files
This commit introduces some functions for reading keys and
certificates from PEM files, and dynamically constructing java
KeyStore instances in memory without requiring a .jks file on
disk.

It also introduces some new configuration options that may
be specified in the `jetty` section of the PuppetDB config
to initialize the web server SSL settings based on your
Puppet PEM files.  This provides us with a path forward for
removing the `keytool` calls and such from the `puppet-ssl-setup`
script in the future.

Conflicts:

	README.md
	src/com/puppetlabs/puppetdb/cli/services.clj