Revert extra SSL config options from services
After discussing with Nick Lewis this morning, we decided that the
configuration code for supporting PEM-based SSL auth in our
main services namespace was nasty enough that it's probably not
worth merging it in for now, since we don't yet have a plan for
how we'd allow users to take advantage of it anyway (they'd need
to either make a copy of their puppet pems or set permissions
on them such that they could be read by the PuppetDB user).

We do still want the SSL pem library code, because Ken already
has another use for it relating to Postgres.  I also thought
it would be beneficial to leave this configuration code in the
commit history so that we can pull it back in if we do decide
we want to do so at some point in the future.