Add support for export metadata during export/import
This commit adds code to the export command to cause it to generate
an `export-metadata.json` file.  At the time of this writing, the
contents of that file are simply a map containing the version number
of the `replace catalog` command that should be used during import.

The commit also adds code to the import command to cause it to expect
this metadata file, to parse it, and to use the version number that
it finds therein when submitting catalogs via the `replace catalog`
command.

Conflicts:

	src/com/puppetlabs/puppetdb/cli/import.clj