Get rid of `:document_type` in catalog output
The example JSON catalogs that I was given all included a top-level
key called `document_type`, with a value of `"Catalog"`, so I included
that in my export output.  After discussing with Nick Lewis, it
sounds like that is an artifact of some Puppet PSON implementation
and is not desirable or necessary.  This commit removes it.