Change argument names (and other minor tweaks from pull request review)
I was using `f` as an argument name in several places in this code;
Nick pointed out that in clojure that will generally be assumed to
mean that the argument is a function, rather than a file (which is
what I'd intended).  All of those instances have been changed, and
I fixed another typo that was pointed out in the review.