Add a lazy `all-entries` sequence to the tar archive library
Previously, if a caller wanted to loop over all of the available
entries in a tar archive, they'd have to do some fairly nasty
loop/recur stuff.  This commit adds an `all-entries` function,
which returns a lazy sequence of all of the entries in the archive.
This allows callers to write much more idiomatic clojure to iterate
over the entries in the archive.