Filter out nodes that don't have a catalog during export
Prior to this commit, the export tool was trying to get
catalogs for all the nodes that were returned by the `/nodes`
endpoint.  It is possible that some of those nodes won't have
a catalog, and in those cases, the export tool was crashing.

This commit simply applies a filter to remove nodes whose
catalog timestamp is null, thus pruning them from the rest
of the export process.