Fix SO MANY Ruby 1.8.7 issues in storeconfigs spec
When reading our export tarball back and mapping it to a hash, we were
trying to omit directories, which are indicated by a trailing slash.
Unfortunately, on Ruby 1.8.7, indexing into a string returns the integer
value of the character at that position. So we need to use
filename[-1,1] to tell Ruby to get a 1-character long substring instead.

Also, we had been reusing the variable (filename) containing the
location of our tarball as a block variable. In Ruby 1.8, this clobbers
the original variable outside the block, which caused us to be unable to
read the tarball later.

Finally, we were retrieving the content of the tarball using
String#lines. This returns an Enumerator, which for some reason in Ruby
1.8.7 won't implicitly convert to an array when passed as an argument to
zip, so we have to do it manually.