[packaging] (#19326) terminus files conflict
Prior to commit b2aa562, the terminus package simply owned
%{puppet_libdir}/puppet, which conflicted with the puppet package. That commit
changed the terminus package files to explicitly own every file, thereby
removing the conflict. However, as kbarber points out, the terminus package
files are still owned by (pe-)puppet, which is unnecessary and different than
the puppet package - the files should just be owned by root, both to align
ownership and to minimize any sort of attack surface if the puppet user is
compromised. This commit updates the packaging to change the ownership to root.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>