Import/export acceptance tests
This commit adds two new acceptance tests: one for testing an
export from legacy storeconfigs, and one for testing the new
puppetdb import/export functionality directly.  The test of the
new import/export functionality is passing, but the legacy
storeconfigs is currently failing because PuppetDB can't handle
a catalog submission where there are no edges present in the
catalog.  After discussing this with Nick, we've decided that
the best solution is to alter the legacy export tool to generate
edges for the resources, rather than modifying PuppetDB to support
catalogs that don't contain edges.

This commit also adds some utility methods for recursive hash
diffing (required for comparing the exported JSON catalogs), and
a small fix to add the `rake package:bootstrap` target to the
setup steps if you are trying to run the acceptance tests locally.