Allow override of `event-query-limit` via query parameter
After discussing with Nick and Ruth, we determined that there will
probably be some Burgundy use cases where the PuppetDB-wide
configuration setting for `event-query-limit` may not match up
with the needs of Burgundy.  (e.g., when we add support for
"group by" and filtering the list of returned fields, we should
be able to handle a much larger number or result rows than we
will for the default event query path.

This commit simply allows you to pass in an extra query parameter,
`limit`, which will override the value from the configuration file.