Change `pmap` back to `map` in export tool
Because clojure's `pmap` attempts to evaluate lazy sequences in
chunks of 32, the existing code for the export tool could cause
us to run out of Postgres connections because of too many
simultaneous catalog requests.  This commit simply changes the
code back to using a regular `map`, which is slower but safer
for now.  I may submit another pull request with another implementation
of parallelization if I can put together something simple without
spending too much time on it, but for now, it's best for us to
just make sure that we're safe and stable since performance isn't
a hugely pressing concern for this tool.