Docs: Change canonical links to "latest" instead of 1.1
This should save some time when pushing new versions. I knew that "latest"
shortcut would come in handy.

In our implementation, "latest" points at the newest numeric version, never at
named versions like "master." (Versions, of course, are created in
puppet-docs/source/_config.yml. Apologies as ever for this toolchain, but
anyway.)

The one funky bit in this commit is that I changed one canonical link from 1.2
to latest, which is a bunk URL until we release the 1.2 docs. I am fine with
that, and Google will figure it out until then.

Conflicts:
	documentation/migrate.markdown