Change USING clause to more explicit ON in several SQL queries
HSQL was doing something really, really stupid when we tried to
use a USING clause before multiple INNER JOINs.  This was causing
`get-edges` to return multiple copies of edges for any node
that had more than one catalog in the database.  Changing to a
more explicit ON clause fixes the problem, so we went ahead and
made this replacement everywhere in the code where we had previously
been using USING.

This also forced us to be more explicit about some of the fields
that we refer to in our SELECT statements.