Port event query tests over to use event `:test-id`s
The path I was going down where I was trying to filter the example
events directly in clojure in order to test them against the real
ones returned by the query was starting to seem like more trouble
than it was worth, since the tests had to validate that my logic
was correct in addition to validating that they query results
were correct.  This commit ports the tests over to having explicit,
hard-coded id's for the events from the example data, and they can
simply validate the query response against those rather than trying
to compute them.