Ditch custom JSON streaming func
Turns out that I was wrong, and cheshire already implements the same
algorithm I was using in `stream-json`. No reason to reinvent the wheel.
I left the function in place as an abstraction over the specific
parameters we'd like to pass to cheshire (useful once we start streaming
other things beyond resources).