Remove certnames from exported reports
As a side effect of a recent merge, I introduced a bug in the
report export functionality.  Report submissions should not
contain the 'certname' field on individual resource events,
but that field does come back in query results now.  This commit
simply removes it from the export data so that we don't try to
send it back up on imports.