Remove parallelization during export
In previous commits, we attempted to limit resource consumption
by using a "bounded" pmap when parallelizing our data exports.
This works great for preventing pmap from executing too many
operations simultaneously, but it does not prevent pmap from
building up a large backlog of work in the seq that it is
constructing; therefore, if the consumer of the seq is not
processing the work quickly enough, we can still end up using
a ton of memory.

The long-term solution for this is going to be to use a
`BlockingQueue` to prevent the pmap from generating work too
quickly.  However, the implementation of that can be a bit
tricky so for now we're opting for stability over performance.

This commit simply removes the parallelization and handles the
export serially.

In the future what I'd like to do is to polish up the
producer/consumer library that I wrote for this, and then use
that to solve the problem on both the import and export sides.