(#21321) Configurable SSL protocols
This patch adds an additional configuration option, `ssl-protocols`, to
the `[jetty]` section of the configuration file. This lets users specify
the exact list of SSL protocols they wish to support, such as in cases
where they're running PuppetDB in an environment with strict standards
for SSL usage.

If the option is not supplied, we use the default set of protocols
enabled by the local JVM.