Clean up test logging for mq-related tests
Prior to this commit, many of the tests that were related to
ActiveMQ were writing a ton of distracting log output
(including stack traces) to the console.  This made it harder
to identify "real" failures in the test output.

This commit adds some more robust logic to the `with-log-output`
macro, which redirects log messages to an atom.  It also creates
a fixture that can be used to enable this macro on every test
in a file, and uses that fixture for the mq tests.

End result: no more false alarms from the mq tests.