(#21614) Include /v2 in metrics end-point examples
Previously we weren't adding the /v2 to the examples show in this document.
While it is true there is no difference between v1 & v2 in this case, we
should keep the convention to avoid confusion, and to make clear the end-point
is part of the versioned structure.

Signed-off-by: Ken Barber <ken@bob.sh>
(cherry picked from commit d38d38eac3055a9711cd33047271612ae877899a)

Conflicts:
	documentation/api/query/v2/metrics.markdown