(#21611) Allow rake commands to be ran on Ruby 2.0
This allows rake commands to be ran on Ruby 2.0, for building on Fedora 19
to be made possible. This also removes the current behaviour whereby
unsupported rubies force the command to exit, instead we now just return
a warning to STDERR and keep going in the hope that things will work anyway,
so we don't block people that insist on using the latest Ruby before we
are ready.

Signed-off-by: Ken Barber <ken@bob.sh>