(maint) Get PE-PuppetDB Acceptance Tests Passing
Prior to this commit, the PE-PuppetDB Acceptance tests were not passing
on the Jenkins job running on jenkins-enterprise. This commit fixes that
in several different fixes mentioned below:

Justins fixes:

- use Puppet Enterprise specific options file since "common" includes FOSS setup suite
- use bash, pe-postgres doenst otherwise have a shell

My Fixes:

- Changing username/pass for database
  - This is a temp solution. A better solution would be to get these
    values from the harness and set dynamically
- Add quotes to password setting
- Test was using wrong hostname. Changing it to rechable_name will fix this problem.

(maint) Change how cert_whitelist deals with files

Prior to this commit, the whitelist was not being used correctly and the
test was failing. This commit hopefully fixes that issue by changing how
grep is used.