(maint) Move suse specific java requirements into the open_jdk macro
There is an open_jdk macro defined at the top of the spec file to specify the
java requirements for puppetdb. This commit moves the sles/suse specific java
logic into that macro section. This supplies the os standard java for opensuse
11,12 and sles. java-devel should not be required for puppetdb build or
installation, so this commit also removes that dependency.