Modify Rakefile to support Ruby 2.0
Prior to this commit, there was no support for Fedora 19. This
commit adds that by including the f19 mock in the build_defaults
file, as well as by editing the Rakefile to allow for Ruby 2.0,
not just ruby 1.8 or 1.9. Additionally, `@plibdir` was changed
so that we can assign this path dynamically, rather than statically.