(packaging) Add suse init script to PuppetDB
This commit adds a simple init script for Suse/SLES. This init script is
templated and will function on PE or FOSS. It lays down a pidfile and handles
creating the pidfile directory. It returns LSB compliant exit codes.
It uses pgrep to get the pidfile of the daemon, so a dependency on procps is
added for sles.