Add multitime! macro
Let's say you'd like to execute some code, but you want to wrap it in a
call to `time!` so that its performance is tracked. That's great, but
what if you want to have its performance tracked in multiple timers
simultaneously (e.g. a more specific timer, and a roll-up timer)?

The `multitime!` macro works exactly like `time!`, except it support
multiple timers.