Fix spec tests to be compatible with puppet 3.3
One of our spec tests mocks an object from the puppet object model.
In the latest version of puppet this object had a small API change;
our mock object was not set up to handle this correctly. This
commit should make the tests compatible with all versions of Puppet
from 2.7.x, 3.0.x, 3.1.x, 3.2.x, and 3.3.x.