(maint) puppedb-ssl-setup tests during package upgrade fail
This patch performs different tests around the puppetdb-ssl-setup behaviour
specifically during package upgrades. This is because the legacy configuration
will exist in such an upgrade.

This does not however apply to plain installs, nor does it apply to source
based upgrades since the jetty.ini is overwritten anyway.

These tests will need to be removed when we make a new feature release, as
the behaviour will change again :-(.

Signed-off-by: Ken Barber <ken@bob.sh>