(maint) Refactor Code
This commit refactors the code addressed in my pull request number 593.

In dynamic-query-load.js

Removed repeat of functions and created a single function HTML tags can
call to be able to generate the correct dropdowns and text with the
onclick method. The function requires the version and endpoint number,
key associated with the QUERY_MAP constant, and if valid a start and end
range for what the tag wants highlighted from the loaded in query. This
change also adds a QUERY_MAP hashmap constant.

In dashboard.js -> query-dashboard.js

Created two constants: The name of the cookie saved in the users browser
for previous queries, and an endpoint map with the versions as the key,
  and the endpoints as the value. The function getMenuNumber now goes
  through the END_POINTS_MAP to figure out which menu number it relates
  to in the dropdowns. It returns the number it finds. isTextboxVisible
  and isDropdownVisible are changed to update because they do not return
  true or false, but change the visibility of the textbox/dropdowns.

In query.html

The dropdown divs are renamed to show that they are containers, not the
dropdown tag itself. All example query tags have changed their function
to match the single one created in dynamic-query-load.js.