(PE-962) Update latest report for node when storing reports
Prior to this commit, we had the new schema and migration
code for flagging the latest report for each node, but
we weren't updating that data when new reports came in.

This commit adds the code to do those updates every time
a new report is stored, and tests to verify that this
behavior works as expected.