Remove unused settings relating to queued commands
Some time ago, we did some investigatory work around the possibility
of queueing failed commands to disk on the master and trying to
re-send them to puppetdb later.  This work ended up being scrapped
due to complications relating to the unpredictable threading/process
model for the environments that the puppet master might be
deployed in.  However, it appears that we failed to remove one of
the configuration settings that had been added in support of that.