(#22122) Fix ssl-setup error when no puppet certs exist
In the case where puppetdb-ssl-setup is ran without certificates pre-existing
the script current just returns an unhelpful error.

This was bad, as this often occurs during package installation, leaving the
user with a feeling that the package installation is now broken.

This patch means the tool now returns a meaningful error, and tries to advise
the user as to what to do next.

I've added coverage specially for this case as well.

Signed-off-by: Ken Barber <ken@bob.sh>