(PE-957) Filter out default "skipped" Schedule events
In versions of puppet prior to 3.3, every single agent run
will generate a swath of skipped Schedule events, which
are not useful and clutter up the reports data in PuppetDB.

This commit adds some "blacklist" functionality, whereby
certain well-known events can be filtered in the PuppetDB
report processor before being sent up to PuppetDB, so
they won't be stored.

There is a configuration option that can be used to
disable the blacklist, just in case these events are meaningful
to some set of end-users.

The blacklist is hard-coded for now and only contains the list
of skipped Schedule events that are well-known.