(Bug #22165) rm hardcoded hostname dependencies
Prior to this commit, we had hardcoded burji.puppetlabs.com in the
build_defaults file. We've come to the conclusion that this isn't the
best way thing to do. This commit takes out those hardcoded hostnames,
in favor of cnames that point to the right place, and will be easier to
update next time we change things up.