(PE-957) Fix spec so that it doesn't assume array order
One of the tests for filtering blacklisted events was
assuming that the array of events returned by `report_to_hash`
would always be in the same order; in fact, the order
differed between ruby 1.8.7 and ruby 1.9.3.  This commit
changes the test so that it doesn't assume a specific order,
since the order doesn't really have anything to do with
what we're testing.