(maint) Code Refactor Improvement
This commit changes a lot of the code based on the code review done on
Github. A lot of improvements include removing some attributes in the js
script html tag that are no longer required. Cleaning up some of the
code from the original dashboard. Replacing all of the
'document.getElementById' with the JQuery version to make the code
cleaner. Moving the PuppetDB vesion js from index.html to use the js
file query.html uses (to help reduce duplicate code). Changing the
error messages to use the bootstrap alert div instead of a font tag.
Abstracting some of the bootstrap css that was wthin the html files
into its own css.