(#22168) Adjust max-threads based on cpu
This patch fixes a weird behaviour in Jetty 7 whereby if you attempt to set the
max-threads to something lower then the number of CPU's on your system, it will
simply stop with the warning:

  insufficient threads configured for SslSelectChannelConnector@0.0.0.0:8081

At this point it will simply not serve any HTTP request at all, effectively
breaking PuppetDB.

This also applies to the default setting of 50. So users with greater then 50
cpus will always see this error, and may not know how to resolve it themselves.

This patch adjusts the behaviour so we automatically adjust the # of
max-threads to a safe level if below num-cpus, and we log a warning telling the
user of this fact.

Signed-off-by: Ken Barber <ken@bob.sh>