Add transaction-uuid key to catalog if it doesn't exist
Since old versions of puppet don't have a transaction-uuid
in their catalog, and since we need to support imports of data
that was exported from prior versions of puppetdb, the
catalog transformation logic needs to tolerate the
absence of this key.

This commit simply adds the key (with a value of `nil`) if the
key isn't already present.