Add strict query param validation for all endpoints
This commit gets rid of the specific query param presence/absence
validation that we were doing for the generic paging query parameters,
and replaces it with a more general white-list approach.  All query
endpoint now specify a list of required/optional query params, and
these are validated via the middleware.