Change return value of `paged-query-to-vec` to a map
`paged-query-to-vec` needs to be able to return the total count
of rows that would have been returned by a paged query, in
addition to the actual result set.  In order to accomplish this,
we need to change the return type from an array to a map.

For now, this map contains just a single key, `:results`, whose
value is the same vector of results that the function was previously
returning.  In a subsequent commit we'll add a `:count` (or similar)
key that can be optionally populated with the count data.