Update docstrings and docs re: paged queries
This commit updates the docstrings of some of the functions in
the `query` namespace which used to return a vector of query
results but now have a slightly more complex return type.

This also updates the end-user documentation to indicate how to
use the new `count?` query parameter to request result counts
for paged queries.