Rename `:results` key in query result map to `:result`
After discussion, we decided that `:result` was a better name for the
key in the query map because in the general case there is no reason
to assume that the result must always be a collection.  This
makes it a bit more obvious that we can write general functions,
like the `query-result-response` function in the `puppetdb.http`
namespace, which can do things like simply convert the `:result`
to JSON without being concerned about the specific details of
its contents.