Retroactively add replace-catalog version 3
We introduced a very minor change in the catalog command as part
of the burgundy run-up (new `transaction-uuid` field).  When we
did this, we simply hacked it into v2 of the replace-catalog command
without giving it enough thought.

After further discussion with @nicklewis, I think we've pretty
much decided that we need to stick with our existing precedent
of "any time *anything* changes, bump the corresponding version
number".

This commit goes back in and introduces v3 of "replace catalog",
and refactors things so that v2 goes back to working the
way it would have before.

It also updates some documentation around this, and around
the "store report" command (whose version was bumped around the
same time, but whose documentation had not been updated
accordingly).