tests: add ruby tests to travis-ci
This switches the TravisCI matrix to support different languages so we can do
tests for Clojure and for our Terminus as well for PR's.

For now this just supports the default version of Ruby that travis-ci provides
due to hassles trying to use 'rvm use' in a script. Later on we'll extend the
tooling to support this but for now having at least the default (1.9.3) tested
is going to get us further then where we are now.

Signed-off-by: Ken Barber <ken@bob.sh>