Rename all occurrences of 'node' to 'certname'
We'd originally planned to go through and make the change from
`certname` to `node` across the whole PuppetDB API as part of
this release.  Because of that, we'd started using the word
`node` in the new places in the event-counts/etc. APIs.

It's now looking way too risky to try to make the full
`certname`->`node` change right now, so we're punting that...
but... that means that we have to change the new stuff from
`node` back to `certname` so that the API will at least
be consistent.  This commit does that.